src/online_label_api_dev/requirement.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cymysql==0.9.13
dnspython==1.16.0
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Executor==0.9.2
Flask-HTTPAuth==3.2.4
Flask-SocketIO==4.1.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
futures==3.3.0
greenlet==0.4.15
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
libscrc==0.1.2
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.16.2
pexpect==4.7.0
ptyprocess==0.6.0
python-engineio==3.8.1
python-magic==0.4.15
python-socketio==4.1.0
PyYAML==5.1.1
redis==3.2.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.2
Werkzeug==0.15.2
WTForms==2.2.1
